SF#3563517: Get rid of loss-of-precision warnings on x64 MacOS

+++ Foundation/include/Poco/Thread_POSIX.h
@@ -215,7 +215,7 @@ inline void ThreadImpl::yieldImpl()
 
 inline int ThreadImpl::getStackSizeImpl() const
 {
-	return _pData->stackSize;
+	return static_cast<int>(_pData->stackSize);
 }
